docs(notebook-cloud): remove Access demo path
