+++ docs/architecture/cleanup-punchlist.md
@@ -72,7 +72,7 @@ Severity legend:
 
 | ID | Smell | Severity | Where |
 |----|-------|----------|-------|
-| HCA-1 | Mixed hosted credentials are documented as reject-by-default in `hosted-credential-transport.md`, but the current notebook-cloud prototype still has Access-specific credential precedence. The direct-OIDC implementation should reject mixed identity-bearing credentials unless a deployment proves they are one credential. | Targeted PR | `apps/notebook-cloud/src/identity.ts`, `apps/notebook-cloud/test/identity.test.ts`, `docs/architecture/hosted-direct-oidc-demo-runbook.md` |
+| HCA-1 | Mixed hosted credentials are documented as reject-by-default in `hosted-credential-transport.md`, but the current notebook-cloud prototype still needs complete mixed-credential rejection coverage. The direct-OIDC implementation should reject mixed identity-bearing credentials unless a deployment proves they are one credential. | Targeted PR | `apps/notebook-cloud/src/identity.ts`, `apps/notebook-cloud/test/identity.test.ts`, `docs/architecture/hosted-direct-oidc-demo-runbook.md` |
 | HCA-2 | Capability naming still needs product polish: `editor`/`owner` have only the widget comm-state exception on `RuntimeStateDoc`, while runtime lifecycle/progress/output authority requires explicit `runtime_peer` and execution intent requires the request API. Align UI copy, request-role wording, and authorization helpers so "can edit notebook" never reads as "can author runtime state." | Design | `apps/notebook-cloud/src/authorization.ts`, `apps/notebook-cloud/src/room-materializer.ts`, `crates/nteract-identity/src/lib.rs`, `docs/architecture/hosted-room-authorization.md` |
 | HCA-3 | Editor blob-upload authority is unresolved. The runtime/blob audit allows editor uploads as precursors to authorized references, while current cloud code and README deny editor uploads. Staged policy should be explicit: editors cannot upload until server-side reference-path validation lands, or editor upload and path validation must ship together. | Design | `apps/notebook-cloud/src/identity.ts`, `apps/notebook-cloud/README.md`, `docs/architecture/runtime-peer-and-blob-authority-audit.md` |
 | HCA-4 | Live peer validators must reject incoming changes authored by `system`/legacy schema actors. Schema/system actors are tolerated only as trusted seed or import history already present before peer ingress, not as newly received peer-authored deltas. | Targeted PR | `crates/notebook-doc/src/lib.rs`, `crates/runtime-doc/src/doc.rs`, `docs/architecture/identity-and-trust.md` |
